Changed LottieIsland to be based of command line argument
