+++ Samples/Islands/DrawingIsland/DrawingCsTestApp/Program.cs
@@ -28,38 +28,21 @@ window.Show();
 
 var compositor = new Compositor();
 
-// DrawingIsland
-//var drawing = new DrawingIsland(compositor);
-
-//var siteBridge = DesktopChildSiteBridge.Create(compositor, window.Id);
-//siteBridge.ResizePolicy = ContentSizePolicy.ResizeContentToParentWindow;
-//siteBridge.Show();
-//siteBridge.Connect(drawing.Island);
-
-// LottieIsland
-var lottieIsland = LottieContentIsland.Create(compositor);
-var lottieVisualSource = LottieVisualSourceWinRT.CreateFromString("ms-appx:///Assets/LottieLogo1.json");
-if (lottieVisualSource != null)
-{
-    lottieVisualSource.AnimatedVisualInvalidated += (sender, args) =>
-        {
-            object? diagnostics = null;
-            IAnimatedVisualFrameworkless? animatedVisual = lottieVisualSource.TryCreateAnimatedVisual(compositor, out diagnostics);
-            if (animatedVisual != null && lottieIsland != null)
-            {
-                compositor.DispatcherQueue.TryEnqueue(() =>
-                {
-                    lottieIsland.AnimatedVisual = animatedVisual;
-                    lottieIsland.PlayAsync(0, 1, true);
-                });
-            }
-        };
-}
-
 var siteBridge = DesktopChildSiteBridge.Create(compositor, window.Id);
 siteBridge.ResizePolicy = ContentSizePolicy.ResizeContentToParentWindow;
 siteBridge.Show();
-siteBridge.Connect(lottieIsland.Island);
+
+if (args.Contains("Lottie"))
+{
+    // LottieIsland
+    LottieIslandProgram.CreateLottieIsland(compositor, siteBridge);
+}
+else
+{
+    // DrawingIsland
+    var drawing = new DrawingIsland(compositor);
+    siteBridge.Connect(drawing.Island);
+}
 
 queue.RunEventLoop();
